+++ jmaqs-webservices-jdk11/src/test/java/com/magenic/jmaqs/webservices/jdk11/WebServiceDriverGetUnitTest.java
@@ -23,6 +23,12 @@ public class WebServiceDriverGetUnitTest extends BaseWebServiceTest {
    * String to hold the URL.
    */
   private static final String baseUrl = WebServiceConfig.getWebServiceUri();
+  
+  /**
+   * The web service driver to be used in a test.
+   */
+  private final WebServiceDriver webServiceDriver = new WebServiceDriver(HttpClientFactory.getDefaultClient());
+
 
   /**
    * Test Json Get deserialize a single product.
@@ -31,8 +37,7 @@ public class WebServiceDriverGetUnitTest extends BaseWebServiceTest {
    */
   @Test(groups = TestCategories.WEB_SERVICE)
   public void getProductXmlDeserialize() throws IOException, InterruptedException {
-    WebServiceDriver client = new WebServiceDriver(HttpClientFactory.getDefaultClient());
-    HttpResponse<String> response = client.get(baseUrl + "/api/XML_JSON/GetProduct/2", MediaType.APP_XML, false);
+    HttpResponse<String> response = webServiceDriver.get(baseUrl + "/api/XML_JSON/GetProduct/2", MediaType.APP_XML, false);
     Product products = WebServiceUtilities.getResponseBody(response, MediaType.APP_XML, Product.class);
     Assert.assertEquals(products.getName(), "Yo-yo", "Expected 3 products to be returned");
   }
@@ -44,8 +49,7 @@ public class WebServiceDriverGetUnitTest extends BaseWebServiceTest {
    */
   @Test(groups = TestCategories.WEB_SERVICE)
   public void getProductsXmlDeserialize() throws IOException, InterruptedException {
-    WebServiceDriver client = new WebServiceDriver(HttpClientFactory.getDefaultClient());
-    Product[] products = client.getContent(baseUrl + "/api/XML_JSON/GetAllProducts", MediaType.APP_XML, false, Product[].class);
+    Product[] products = webServiceDriver.getContent(baseUrl + "/api/XML_JSON/GetAllProducts", MediaType.APP_XML, false, Product[].class);
     Assert.assertEquals(products.length, 3, "Expected 3 products to be returned");
   }
 
@@ -56,8 +60,7 @@ public class WebServiceDriverGetUnitTest extends BaseWebServiceTest {
    */
   @Test(groups = TestCategories.WEB_SERVICE)
   public void getProductJsonDeserialize() throws IOException, InterruptedException {
-    WebServiceDriver client = new WebServiceDriver(HttpClientFactory.getDefaultClient());
-    HttpResponse<String> response = client.getContent(baseUrl + "/api/XML_JSON/GetProduct/2", MediaType.APP_JSON, false);
+    HttpResponse<String> response = webServiceDriver.getContent(baseUrl + "/api/XML_JSON/GetProduct/2", MediaType.APP_JSON, false);
     Product products = WebServiceUtilities.getResponseBody(response, MediaType.APP_JSON, Product.class);
     Assert.assertEquals(products.getName(),"Yo-yo", "Expected 3 products to be returned");
   }
@@ -69,8 +72,7 @@ public class WebServiceDriverGetUnitTest extends BaseWebServiceTest {
    */
   @Test(groups = TestCategories.WEB_SERVICE)
   public void getProductsJsonDeserialize() throws IOException, InterruptedException {
-    WebServiceDriver client = new WebServiceDriver(HttpClientFactory.getDefaultClient());
-    Product[] products = client.get(baseUrl + "/api/XML_JSON/GetAllProducts", MediaType.APP_JSON, false, Product[].class);
+    Product[] products = webServiceDriver.get(baseUrl + "/api/XML_JSON/GetAllProducts", MediaType.APP_JSON, false, Product[].class);
     Assert.assertEquals(products.length, 3, "Expected 3 products to be returned");
   }
 
@@ -81,8 +83,7 @@ public class WebServiceDriverGetUnitTest extends BaseWebServiceTest {
    */
   @Test(groups = TestCategories.WEB_SERVICE)
   public void getProductsPlainText() throws IOException, InterruptedException {
-    WebServiceDriver client = new WebServiceDriver(HttpClientFactory.getDefaultClient());
-    HttpResponse<String> result = client.get(baseUrl + "/api/String/Get",  MediaType.PLAIN_TEXT, false);
+    HttpResponse<String> result = webServiceDriver.get(baseUrl + "/api/String/Get",  MediaType.PLAIN_TEXT, false);
     Assert.assertTrue(result.body().contains("Tomato Soup"),
         "Was expecting a result with Tomato Soup but instead got - " + result);
     Assert.assertTrue(result.body().contains("Yo-yo"),
@@ -98,8 +99,7 @@ public class WebServiceDriverGetUnitTest extends BaseWebServiceTest {
    */
   @Test(groups = TestCategories.WEB_SERVICE)
   public void getStringById() throws IOException, InterruptedException {
-    WebServiceDriver client = new WebServiceDriver(HttpClientFactory.getDefaultClient());
-    HttpResponse<String> result = client.get(baseUrl + "/api/String/1",  MediaType.PLAIN_TEXT, false);
+    HttpResponse<String> result = webServiceDriver.get(baseUrl + "/api/String/1",  MediaType.PLAIN_TEXT, false);
     Assert.assertTrue(result.body().contains("Tomato Soup"),
         "Was expecting a result with Tomato Soup but instead got - " + result);
   }
@@ -111,8 +111,7 @@ public class WebServiceDriverGetUnitTest extends BaseWebServiceTest {
    */
   @Test(groups = TestCategories.WEB_SERVICE)
   public void getStringByName() throws IOException, InterruptedException {
-    WebServiceDriver client = new WebServiceDriver(HttpClientFactory.getDefaultClient());
-    HttpResponse<String> result = client.get(baseUrl + "/api/String/Yo-yo",  MediaType.PLAIN_TEXT, false);
+    HttpResponse<String> result = webServiceDriver.get(baseUrl + "/api/String/Yo-yo",  MediaType.PLAIN_TEXT, false);
     Assert.assertTrue(result.body().contains("Yo-yo"),
         "Was expecting a result with Yo-yo but instead got - " + result);
   }
@@ -123,10 +122,8 @@ public class WebServiceDriverGetUnitTest extends BaseWebServiceTest {
    * @throws InterruptedException if exception is thrown
    */
   @Test(groups = TestCategories.WEB_SERVICE)
-  public void getResponseAndDeserializeJson()
-      throws IOException, InterruptedException {
-    WebServiceDriver client = new WebServiceDriver(HttpClientFactory.getDefaultClient());
-    HttpResponse<String> message = client.get(baseUrl + "/api/XML_JSON/GetAllProducts",
+  public void getResponseAndDeserializeJson() throws IOException, InterruptedException {
+    HttpResponse<String> message = webServiceDriver.get(baseUrl + "/api/XML_JSON/GetAllProducts",
        MediaType.APP_JSON, true  );
     Product[] products = WebServiceUtilities.deserializeJson(message, Product[].class);
     Assert.assertNotNull(products);
@@ -139,10 +136,8 @@ public class WebServiceDriverGetUnitTest extends BaseWebServiceTest {
    * @throws InterruptedException if exception is thrown
    */
   @Test(groups = TestCategories.WEB_SERVICE)
-  public void getResponseAndDeserializeXml()
-      throws IOException, InterruptedException {
-    WebServiceDriver client = new WebServiceDriver(HttpClientFactory.getDefaultClient());
-    HttpResponse<String> message = client.get(baseUrl + "/api/XML_JSON/GetAllProducts",
+  public void getResponseAndDeserializeXml() throws IOException, InterruptedException {
+    HttpResponse<String> message = webServiceDriver.get(baseUrl + "/api/XML_JSON/GetAllProducts",
         MediaType.APP_XML, true);
     Product[] products = WebServiceUtilities.deserializeXml(message, Product[].class);
     Assert.assertNotNull(products);
@@ -155,11 +150,9 @@ public class WebServiceDriverGetUnitTest extends BaseWebServiceTest {
    * @throws InterruptedException if exception is thrown
    */
   @Test(groups = TestCategories.WEB_SERVICE)
-  public void getResponseAndDeserializeJsonExpectedStatus()
-      throws IOException, InterruptedException {
-    WebServiceDriver client = new WebServiceDriver(HttpClientFactory.getDefaultClient());
-    HttpResponse<String> message = client.get(baseUrl + "/api/XML_JSON/GetAllProducts",
-        MediaType.APP_JSON, HttpStatus.OK);
+  public void getResponseAndDeserializeJsonExpectedStatus() throws IOException, InterruptedException {
+    HttpResponse<String> message = webServiceDriver.get(
+        baseUrl + "/api/XML_JSON/GetAllProducts", MediaType.APP_JSON, HttpStatus.OK);
     Product[] products = WebServiceUtilities.deserializeJson(message, Product[].class);
     Assert.assertNotNull(products);
     Assert.assertEquals(products.length, 3, "Expected 3 products to be returned");
@@ -171,13 +164,47 @@ public class WebServiceDriverGetUnitTest extends BaseWebServiceTest {
    * @throws InterruptedException if exception is thrown
    */
   @Test(groups = TestCategories.WEB_SERVICE)
-  public void getResponseAndDeserializeXmlExpectedStatus()
-      throws IOException, InterruptedException {
-    WebServiceDriver client = new WebServiceDriver(HttpClientFactory.getDefaultClient());
-    HttpResponse<String> message = client.get(baseUrl + "/api/XML_JSON/GetAllProducts",
-        MediaType.APP_XML, HttpStatus.OK);
+  public void getResponseAndDeserializeXmlExpectedStatus() throws IOException, InterruptedException {
+    HttpResponse<String> message = webServiceDriver.get(
+        baseUrl + "/api/XML_JSON/GetAllProducts", MediaType.APP_XML, HttpStatus.OK);
     Product[] products = WebServiceUtilities.deserializeXml(message, Product[].class);
     Assert.assertNotNull(products);
     Assert.assertEquals(products.length,3,"Expected 3 products to be returned");
+  }
+
+  /**
+   * Test type parameterized Get request with expected status.
+   * @throws IOException if an exception is thrown
+   * @throws InterruptedException if an exception is thrown
+   */
+  @Test(groups = TestCategories.WEB_SERVICE)
+  public void getTypeParamWithExpectedStatus() throws IOException, InterruptedException {
+    Product[] res = webServiceDriver.get(baseUrl + "/api/XML_JSON/GetAllProducts",
+        MediaType.APP_JSON, HttpStatus.OK, Product[].class);
+    Assert.assertNotNull(res);
+  }
+
+  /**
+   * Test Get request with expected status.
+   * @throws IOException if an exception is thrown
+   * @throws InterruptedException if an exception is thrown
+   */
+  @Test(groups = TestCategories.WEB_SERVICE)
+  public void getWithExpectedStatus() throws IOException, InterruptedException {
+    HttpResponse<String> res = webServiceDriver.get(
+        baseUrl + "/api/XML_JSON/GetAllProducts", MediaType.APP_XML, HttpStatus.OK);
+    Assert.assertNotNull(res);
+  }
+
+  /**
+   * Test Get with response request with expected status.
+   * @throws IOException if an exception is thrown
+   * @throws InterruptedException if an exception is thrown
+   */
+  @Test(groups = TestCategories.WEB_SERVICE)
+  public void getWithResponseWithExpectedStatus() throws IOException, InterruptedException {
+    HttpResponse<String> res = webServiceDriver.get(
+        baseUrl + "/api/XML_JSON/GetAllProducts", MediaType.APP_XML, HttpStatus.OK);
+    Assert.assertNotNull(res);
   }
 }
